+++ mitra_activitypub/src/builders/update_note.rs
@@ -40,7 +40,6 @@ fn build_update_note(
     instance_url: &str,
     media_server: &MediaServer,
     post: &Post,
-    fep_e232_enabled: bool,
 ) -> UpdateNote {
     let authority = Authority::server(instance_url);
     let object = build_note(
@@ -49,7 +48,6 @@ fn build_update_note(
         &authority,
         media_server,
         post,
-        fep_e232_enabled,
         false, // no context
     );
     let primary_audience = object.to.clone();
@@ -73,7 +71,6 @@ pub async fn prepare_update_note(
     media_server: &MediaServer,
     author: &User,
     post: &Post,
-    fep_e232_enabled: bool,
 ) -> Result<OutgoingActivityJobData, DatabaseError> {
     assert_eq!(author.id, post.author.id);
     let activity = build_update_note(
@@ -81,7 +78,6 @@ pub async fn prepare_update_note(
         &instance.url(),
         media_server,
         post,
-        fep_e232_enabled,
     );
     let recipients = get_note_recipients(db_client, post).await?;
     Ok(OutgoingActivityJobData::new(
@@ -120,7 +116,6 @@ mod tests {
             instance_url,
             &media_server,
             &post,
-            false,
         );
 
         assert_eq!(activity.id.starts_with(instance_url), true);